Format api for basic steps as table
This makes things easier to read, since it's more of a usage example.

The commit also moves all references to the upstream documentation down to the end of the document. That way, when we bump to a new upstream release, we will likely only have to modify that part of the file.