Uptate option handling in TypeScript
Having overly specific types can make it hard to write some more generic
option-writing code.  This commit attempts to establish some kind of best
practice: Use the dict when you know the name of the option, and ise the
optGet and optSet method if that name is not a literal but a variable.  That
way a variable which may be an arbitrary string can still get type-checked
correctly, although it may of course throw an exception at runtime.  In a
similar spirit, always use the canonical representation of values even when
setting them.  Other notation is possible, but expressing e.g.  both the
number and the string literal notation for an enum means that its values
couldn't get type-checked as a string any more.