Provide command line interface and building blocks for promise pipelines
All the new API is currently undocumented and experimental.
The command line tool takes tidy-html5/console/tidy.c as its template,
trying to be feature compatible in the features it supports.
Currently there are likely to be very many features which can be configured
but which get ignored at runtime, e.g. those for setting error output file.
The pipeline-building stages may end up with other names, or in a submodule,
or not public at all.  It's also unsure whether the API between them should
be documented or kept opaque.